Revert "Merge remote-tracking branch 'origin/CorpChatMessage' into CorpChatMessage"
This reverts commit 21480322a4075142310f0ae7c6da3507251898d6, reversing
changes made to 432705eee38111d05495847ae05660ae2f56dcab.

+++ runelite-client/src/main/java/net/runelite/client/plugins/xptracker/XpStateSingle.java
@@ -115,11 +115,11 @@ class XpStateSingle
 		return Integer.MAX_VALUE;
 	}
 
-	private double getSkillProgress()
+	private int getSkillProgress()
 	{
 		double xpGained = getCurrentXp() - startLevelExp;
 		double xpGoal = endLevelExp - startLevelExp;
-		return (xpGained / xpGoal) * 100;
+		return (int) ((xpGained / xpGoal) * 100);
 	}
 
 	private String getTimeTillLevel()
